first pass at group API
