+++ lib/tapjoy/ldap.rb
@@ -3,7 +3,6 @@ require 'yaml'
 require 'trollop'
 require_relative 'ldap/cli'
 require_relative 'ldap/base'
-require_relative 'ldap/group'
 require_relative 'ldap/key'
 require_relative 'ldap/audit'
 require_relative 'ldap/version'
